ComponentExamples: Refactor into separate example files
